přidány výjimky a testy je třeba vyřešit chybu s mockito, a oprava

+++ src/main/java/com/example/restapi/todo/TodoService.java
@@ -1,5 +1,6 @@
 package com.example.restapi.todo;
 
+import com.example.restapi.todo.exception.TodoNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -30,26 +31,34 @@ public class TodoService {
     }
 
     public void deleteTodo(Long id) {
+        if(!todoRepository.existsById(id)) {
+            throw new TodoNotFoundException(
+                    "Todo with id " + id + " does not exist");
+        }
+
         todoRepository.deleteById(id);
         todoRepository.flush();
     }
 
     public void updateTodo(Long id, TodoRequest todoRequest) {
-        Optional<Todo> byId = todoRepository.findById(id);
-        if (byId.isPresent()) {
-            Todo todoById = byId.get();
+        if (!todoRepository.existsById(id)) {
+            throw new TodoNotFoundException(
+                    "Todo with id " + id + " does not exist");
+        }
 
-            Todo todo = Todo.builder()
-                    .id(todoById.getId())
-                    .title(todoRequest.getTitle())
+        Optional<Todo> byId = todoRepository.findById(id);
+        Todo todoById = byId.get();
+
+        Todo todo = Todo.builder()
+                .id(todoById.getId())
+                .title(todoRequest.getTitle())
 //                    .content(todoRequest.getContent())
 //                    .description(todoRequest.getDescription())
 //                    .createdBy(todoRequest.getCreatedBy())
 //                    .state(todoRequest.getState())
 //                    .projectId(todoRequest.getProjectId())
-                    .build();
+                .build();
 
             Todo newTodo = todoRepository.saveAndFlush(todo);
-        }
     }
 }
